MEED-345 : Refund option with Metamask account in a drawer .  (#202)
prior to this change, the refunding process was experienced throughout a modal and was only tolerant to the internal wallet, this change has enabled metamask refunding option throughout exo-drawer with alerting system .
