+++ perk-store-services/src/main/resources/locale/addon/PerkStore_en.properties
@@ -71,7 +71,7 @@ exoplatform.perkstore.warning.recipientDoesntHaveWallet=The recipient doesn't ha
 exoplatform.perkstore.label.quantity=Quantity
 exoplatform.perkstore.label.yourOrders=Your orders
 exoplatform.perkstore.label.availableQuantity=Available quantity
-exoplatform.perkstore.label.quantityWithMax=Quantity (max: {0})
+exoplatform.perkstore.label.quantityWithMax=max: {0}
 exoplatform.perkstore.label.quantityPlaceholder=Select a quantity to buy
 exoplatform.perkstore.label.walletPassword=Wallet Password
 exoplatform.perkstore.label.wallet=Wallet
